Categorical Analysis of AIM Data

This function takes the outputs from the functions weight() optionally passed through weight.adjust(), benchmarker(), and a TerrADat SpatialPointsDataFrame with a field for reporting units. It applies spsurvey::cat.analysis() and returns the results

analyze(evaluated.points, point.weights,
  default.reportingunit = "No reporting unit", reportingunit.type = NA,
  adjustedweights = FALSE, conf = 80)

Arguments

evaluated.points
Data frame output from benchmark().

point.weights
Data frame output from weight(), equivalent to weight()[["point.weights"]] or weight()[[2]].

default.reportingunit
A string to populate the reporting unit field with in the case that designs were not restricted by reporting unit. Defaults to "No reporting unit"

reportingunit.type
A string to populate the reporting unit type field with. Depends on the reporting units used, e.g. it may be "Watershed" or "Sage-grouse Habitat". Defaults to NA.

adjustedweights
Logical. If TRUE, use the values in the field ADJWGT instead of the field WGT. Allows for quick comparison between the results of analysis with and without the weight adjustment. Defaults to FALSE.

conf
Numeric. The confidence level. Defaults to 80.

Value

A data frame with the output from spsurvey::cat.analysis()

Examples

analyze()

#> Error in toupper(names(evaluated.points)): argument "evaluated.points" is missing, with no default