Apply project tracking Excel files to imported Design Databases.

Imports plot tracking worksheets used during an AIM project and uses them to assigns statuses to imported Design Databases.

apply.tracking(filenames, path = "", dd.list, dd.names = c(""), tdat,
  target.values = c("Target Sampled", "TS"), deleteoverdraw = T)

Arguments

filenames
A character vector of the filenames (including extension) of the project tracking Excel files to import. If not using the path argument, the filename should include the entire filepath.

path
Optional string specifying a common filepath containing the project tracking sheets to read in. This will be prepended to the values in filenames. If the tracking sheets are in different folder paths, do not provide this.

dd.list
Output from read.dd().

dd.names
An optional character string vector of Design Database names from dd.list to compare against the plot tracking Excel files. If not provided, all of the Design Databases represented in dd.list will be compared and updated.

tdat
Output from read.tdat().

target.values
Character string or character vector. This defines what values in the point fate field count as target points. The function always looks for "Target Sampled" and "TS", so this argument is only necessary if there are additional values in the sample design databases. This is case insensitive

deleteoverdraw
Logical. If TRUE then unsampled overdraw points will be dropped. Defaults to TRUE.

Value

Returns dd.list with sampling dates and statuses added from the plot tracking Excel files.