Quickly drop observations from a data frame based on the values of a single variable

This wrapper for dplyr::filter() will take a data frame (or Spatial Data Frame if the package spdplyr is installed) and remove all observations where the given variable meets the value of the argument dropvalue.

drop.values(df, variable = "", dropvalue = NA, ignore.case = TRUE)

Arguments

df
A data frame or, if spdplyr is installed, spatial data frame to manipulate.

variable
A character string specifying the name of the variable to base the filtering on.

dropvalue
The value to drop observations based on. Can be either a regular expression as a character string to be passed to grepl() or NA. All observations where the variable variable return TRUE when checked against this will be dropped. Defaults to NA.

ignore.case
Logical. If dropvalue is a regular expression, then this argument is passed to grepl() to decide if the search is case sensitive or not. Defaults to TRUE.

Value

The data frame df without the observations where variable matched dropvalue.