Filtering data frames by dates

Given a date field and a lower date, an upper date, or both, filter data.

# S3 method for by.date
filter(data, date.field, after = NULL, before = NULL,
  ignore.year = FALSE)

Arguments

data
Data frame. So long as [] will work on it, it's fine so spatial data frames are valid.

date.field
Character string. The name of the column/variable in the data frame containing the date values. The date values must be interpretable by lubridate::as_date().

after
Optional character string expressing a date in the format YYYY-MM-DD (or MM-DD if ignore.year = TRUE). The earliest date (inclusive) to return data from. Rows containing earlier dates will not be returned.

before
Optional character string expressing a date in the format YYYY-MM-DD (or MM-DD if ignore.year = TRUE). The latest date (inclusive) to return data from. Rows containing later dates will not be returned.

ignore.year
Logical. If TRUE then years will be ignored in comparisons, e.g. If after = "1986-10-30" or after = "10-30" then all rows containing dates from October 30 to December 31 would be returned, regardless of the year. Defaults to FALSE.