Erase a Spatial Polygons Data Frame from another, either with rgeos or ArcPy

Spatial manipulations in R can get, for lack of a better term, squirrelly. Even with multiple failsafes in place to try to compensate for sliver geometry, rgeos::gDifference() still isn't as robust as one might hope. This wrapper removes the geometry of one SPDF from another via a system call to Python using the library ArcPy or via rgeos::gDifference().

flex.erase(spdf, spdf.erase, method = "arcpy", temp.path = "",
  python.search.path = "C:/Python27", sliverdrop = T, sliverwarn = T,
  sliverthreshold = 0.01)

Arguments

spdf
A Spatial Polygons Data Frame to remove FROM.

spdf.erase
A Spatial Polygons Data Frame to remove geometry from spdf WITH.

method
Character string. This must either be "arcpy" or "rgeos" and determines which approach will be used to frames from one another if combine is TRUE. If "arcpy" is used, then R must have write permissions to the folder temp.path and a valid install of ArcPy. This is preferable to "rgeos" because the functions involved tend to crash at random when handling very small remainder geometries. Case insensitive. Defaults to "arcpy".

temp.path
Optional character string. If erase is "arcpy" this must be the path to a folder that R has write permissions to so that a subfolder called arcpy_temp can be created and used for ArcPy erasure steps.

python.search.path
Character string. The filepath for the folder containing pythonw.exe. Defaults to "C:/Python27".

sliverdrop
Optional logical value. If erase is "rgeos" this will be passed to rgeos::set_RGEOS_dropSlivers() to temporarily set the environment during the erasure attempt. Defaults to TRUE.

sliverwarn
Optional logical value. If erase is "rgeos" this will be passed to rgeos::set_RGEOS_warnSlivers() to temporarily set the environment during the erasure attempt. Defaults to TRUE.

sliverdrop
Optional numeric value. If erase is "rgeos" this will be passed to rgeos::set_RGEOS_polyThreshold() to temporarily set the environment during the erasure attempt. Defaults to 0.01.

Value

The remaining geometry and data in spdf after spdf.erase has been removed from it.