Create a SpatialPolygonsDataFrame from the intersection of two SpatialPolygonsDataFrames

Basically a wrapping of rgeos::gIntersection() and raster::intersect() with additional opportunity to call area.add() and automatically added unique identifiers for each resultant polygon

intersect(spdf1, spdf1.attributefieldname,
  spdf1.attributefieldname.output = NULL, spdf2, spdf2.attributefieldname,
  spdf2.attributefieldname.output = NULL, method = "gintersection",
  area.ha = TRUE, area.sqkm = TRUE,
  projection = CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"))

Arguments

spdf1
A SpatialPolygonsDataFrame to be intersected

spdf1.attributefieldname
Name of the field in spdf1 unique to the unit groups or units to take values from.

spdf1.attributefieldname.output
Optional string for the name of the field in the output SPDF to duplicate values from spdf1.attributefieldname into. If unspecified, then spdf1.attributefieldname will be used. Defaults to NULL.

spdf2
A SpatialPolygonsDataFrame to be intersected

spdf2.attributefieldname
Name of the field in spdf2 unique to the unit groups or units to take values from.

spdf2.attributefieldname.output
Optional string for the name of the field in the output SPDF to duplicate values from spdf2.attributefieldname into. If unspecified, then spdf2.attributefieldname will be used. Defaults to NULL.

method
A string specifying which function to use for the interscting step: either rgeos::gIntersection() or raster::intersect(). Valid options are "gintersection" and "intersect". Defaults to "gintersection".

area.ha
Logical. If TRUE, areas will be calculated and added in hectares. Default is TRUE.

area.sqkm
Logical. If TRUE, areas will be calculated and added in square kilometers. Default is TRUE.

projection
An sp::CRS() argument. The final output will be reprojected using this. Defaults to CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs")

Value

A SpatialPolygonsDataFrame with the attributes inherited from spdf1 and spdf2, areas as appropriate, and a unique identifier.

Examples

intersect()

#> Error in intersect(): argument "spdf1" is missing, with no default