Reading in the benchmarks from the Data Explorer

Reading in the benchmarks from the Data Explorer

read.benchmarks(data.path = "", benchmarks.filename = "",
  sheet.name = "Monitoring Objectives", indicator.lut = NULL,
  indicator.lut.benchmarkfield = "indicator.name", convert.l2r = TRUE)

Arguments

data.path
A string specifying the path to the folder containing the .XLSX with the benchmarks

benchmarks.filename
A string specifying the filename of the .XLSX containing the benchmarks

indicator.lut
A data frame with the column "indicator.name" matching the values in the Data Explorer "Indicator" field and one called "indicator.tdat" with corresponding value for the indicators' names in TerrADat.

indicator.lut.benchmarkfield
The name of the column in indicator.lut that matches the "Indicator column of Data Explorer. Defaults to "indicator.name"

Value

A data frame of the benchmarks from the Data Explorer with a field containing an evaluation string to use in testing indicator values against the benchmarks.

Examples

read.benchmarks()

#> Error in sheets_fun(path): zip file '/.xlsx' cannot be opened