Importing Sample Design Databases for AIM Sample Designs

This function imports one or more Sample Design Database[s] and returns a list of the named lists sf, pts, and strata. The named lists contain SpatialPoints/PolygonsDataFrames of the sample frame, points, and strata from each of the geodatabases. The SPDFs are named using the filename of the geodatabase source, so that each list has one SPDF named for each geodatabase imported and those names are identical between lists. If a Sample Design Database is missing any one of those features, a NULL value replaces the SPDF.

read.dd(src = "", dd.src, func = "arcgisbinding",
  target.values = c("Target Sampled", "TS"), omitNAdesignations = FALSE,
  projection = CRS("+proj=longlat +datum=NAD83 +no_defs +ellps=GRS80 +towgs84=0,0,0"))

Arguments

src
Character string defining the filepath containing the sample design database[s].

dd.src
Character string or character vector containing the filenames of the geodatabases to import. Each filename should include the extension ".gdb"

func
Character string. Defines whether to use the rgdal:: or arcgisbinding:: package to read in the geodatabases. Defaults to "arcgisbinding". Valid values are "arcgisbinding" and "readogr". This is not case sensitive

projection
sp::CRS() argument. Defaults to NAD83 with sp::CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"). Is used to reproject all SPDFs in order to perform spatial manipulations

Examples

weighter()

#> Error in weighter(): could not find function "weighter"