Generating summary reports.

Generate both a .PDF and .HTML report with RMarkdown.

report(out.path, project.name, indicator.lut = NULL, benchmarks = NULL,
  analysis = NULL, cats.to.suppress = c("Not Meeting"), point.weights,
  strata.weights, reporting.units.spdf = NULL, sample.frame.spdf,
  project.area.spdf = NULL, points.benchmarked = NULL, dd.points,
  daterange.max = NULL, daterange.min = NULL,
  projection = sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"),
  extension = "html")

Arguments

out.path
Character string. The folder path to write the output reports into.

project.name
Character string. The name of the project to be used in the title of the report and the filenames it's written to.

indicator.lut
Data frame. The lookup table between the indicator names in TerrADat and the human readable indicator names.

benchmarks
Data frame. The output from read.benchmarks().

analysis
Data frame. The data frame "analyses" from the output of analyze().

cats.to.suppress
Character vector. One or more categories to suppress in tables. Defaults to c("Not Meeting").

point.weights
Data frame. The data frame point.weights from the output of weight().

strata.weights
Data frame. The data frame strata.weights from the output of weight().

reporting.units.spdf
Spatial polygons data frame. Used in plotting maps. This MUST have a field named exactly "Type" containing the type of reporting unit (e.g. "Watershed" or "Study Area") and a field named exactly "Subpopulation" which contains the identity of the reporting unit[s] (e.g. the watersheds "Dickshooter Creek" and "Headwaters Deep Creek").

sample.frame.spdf
Spatial polygons data frame. Used in plotting maps.

project.area.spdf
Spatial polygons data frame. Used in plotting maps.

points.benchmarked
Data frame. The output from benchmark().

dd.points
Spatial points data frame or list of spatial points data frames. This should be the pts list in the output from read.dd().

daterange.max
Optional character string. This must be interpretable by lubridate::as_date(), e.g. "2016-04-20". Only sampling locations visited before this date will be considered. Currently only restricts to year.

daterange.min
Optional character string. This must be interpretable by lubridate::as_date(), e.g. "2016-04-20". Only sampling locations visited after this date will be considered. Currently only restricts to year.

projection
Optional sp::CRS() argument. Used to convert points.benchmarked into a spatial points data frame. Only specify if points.benchmarked has coordinates not from the same projection as TerrADat. Defaults to sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0").

extension
Character string. The file extension to write the report to. Defaults to "html".