Calculating Weights from AIM Sample Designs

This function takes the output from the function read.dd() and optionally a SpatialPolygonsDataFrame defining the extent of reporting units. Returns a list of two data frames: strata.weights with the weighting information by stratum and point.weights with the weighting information by point.

weight(dd.import, combine = TRUE, reorder = TRUE, erase = "arcpy",
  temp.path = getwd(), reporting.units.spdf = NULL,
  reportingunitfield = "REPORTING.UNIT", target.values = NULL,
  unknown.values = NULL, nontarget.values = NULL,
  inaccessible.values = NULL, unneeded.values = NULL,
  daterange.max = NULL, daterange.min = NULL,
  fatefieldname = "final_desig", pointstratumfieldname = "dsgn_strtm_nm",
  designstratumfield = "dmnnt_strtm",
  projection = sp::CRS("+proj=longlat +datum=NAD83 +no_defs +ellps=GRS80 +towgs84=0,0,0"),
  sliverdrop = TRUE, sliverwarn = TRUE, sliverthreshold = 0.01)

Arguments

dd.import
Output from read.dd(). This can be the result of running read.dd() on one or many sample design databases

combine
Logical. If provided multiple DDs, should those be combined as part of the weighting calculations? Otherwise the weights will be calculated on a per-DD basis. Defaults to TRUE.

reorder
Logical. If provided multiple DDs, should those be reordered to reflect ascending area? If FALSE, the DDs are considered in the order they appear in dd.import. If combine is FALSE, then this only potentially affects the order of the output information. Defaults to TRUE.

erase
Optional character string. If combine is TRUE this must either be "arcpy" or "rgeos" and determines which approach will be used to frames from one another if combine is TRUE. If "arcpy" is used, then R must have write permissions to the folder temp.path and a valid install of ArcPy. This is preferable to "rgeos" because the functions involved tend to crash at random when handling very small remainder geometries. Defaults to "arcpy".

temp.path
Optional character string. If combine is TRUE and erase is "arcpy" this must be the path to a folder that R has write permissions to so that a subfolder called arcpy_temp can be created and used for ArcPy erasure steps. Defulats to the current working directory.

reporting.units.spdf
SpatialPolygonsDataFrame. Optional reporting units polygons to restrict the sample designs to. If provided, weights will be calculated appropriately based on the intersection of this SPDF and the design[s] Defaults to NULL.

reportingunitfield
Character string. If passing a reporting unit SPDF, what field in it defines the reporting unit[s]?

target.values
Character string or character vector. This defines what values in the point fate field count as target points. The function always looks for "Target Sampled" and "TS", so this argument is only necessary if there are additional values in the sample design databases. This is case insensitive.

unknown.values
Character string or character vector. This defines what values in the point fate field count as unknown points. The function always looks for "Unknown" and "Unk", so this argument is only necessary if there are additional values in the sample design databases. This is case insensitive.

nontarget.values
Character string or character vector. This defines what values in the point fate field count as non-target points. The function always looks for "Non-Target", "NT", and NA, so this argument is only necessary if there are additional values in the sample design databases. This is case insensitive.

inaccessible.values
Character string or character vector. This defines what values in the point fate field count as non-target points. The function always looks for "Inaccessible", so this argument is only necessary if there are additional values in the sample design databases. This is case insensitive.

unneeded.values
Character string or character vector. This defines what values in the point fate field count as not needed or unneeded points. The function always looks for "Not needed", so this argument is only necessary if there are additional values in the sample design databases. This is case insensitive.

daterange.max
Optional character string. This must be interpretable by lubridate::as_date(), e.g. "2016-04-20". Only sampling locations visited before this date will be considered. Currently only restricts to year.

daterange.min
Optional character string. This must be interpretable by lubridate::as_date(), e.g. "2016-04-20". Only sampling locations visited after this date will be considered. Currently only restricts to year.

fatefieldname
Character string defining the field name in the points SPDF[s] in dd.import that contains the point fate. Defaults to "final_desig".

pointstratumfieldname
Character string defining the field name in the points SPDF[s] in dd.import that contains the design stratum. Defaults to "dsgn_strtm_nm".

designstratumfield
Character string defining the field name in the strata SPDF[s] in dd.import that contains the design stratum. Defaults to "dmnt_strtm".

projection
Character string sp::CRS() argument. Defaults to NAD83 with sp::CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"). Is used to reproject all SPDFs in order to perform spatial manipulations.

Examples

weighter()

#> Error in weighter(): could not find function "weighter"