Writing out shapefiles of benchmarked points

Writes out a point shapefile for each management question of points which were classified using the benchmarks for that question. Each

write.benchmarkshp(points.benchmarked, tdat, out.path, name)

Arguments

points.benchmarked
Data frame output from benchmark().

tdat
A Spatial Points Data Frame containing the relevant geometry and fields called PLOTID and PRIMARYKEY.

out.path
A string specifying the output folder path.

name
A string of the project name for use in the filename, e.g. "Idaho_SageGrouse".