Writing out SPDFs as shapefiles

Write our an ESRI shapefile from one or more SPDFs. If multiple SPDFs are being written out, they will be combined, optionally with raster::union() in the case of SpatialPolygonsDataFrame.

write.shapefile(spdf, dd = FALSE, dd.list = "sf", union = TRUE, name,
  type, out.path)

Arguments

spdf
Either a single SPDF, a list of SPDFs, or the output from read.dd(). If multiple SPDFs are provided, they will be output as a single combined shapefile. All SPDFs must be of the same class. If not being combined with a raster::union() or if they are SpatialPointsDataFrames the data slots must have identical variables.

dd
Logical. If the argument spdf is the output from read.dd(), provide TRUE. Defaults to FALSE.

dd.list
A string specifying which of the three lists in dd to pull SPDFs from. Only used if dd = TRUE. Valid values are "sf", "strata", and "pts". Defaults to "sf".

union
Logical. If TRUE and multiple SatialPolygonDataFrames are provided as spdf then raster::union() will be applied.

name
A string of the project name for use in the filename, e.g. "Idaho_SageGrouse".

type
A string of the type of output for use in the filename, e.g. "reportingunits" or "sampleframes".

out.path
A string specifying the output folder path.